+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#2975](https://github.com/oshi/oshi/pull/2975): Query Mac apps in xml formatting using internal xml parsing utils - [@dyorgio](https://github.com/dyorgio).
 * [#3024](https://github.com/oshi/oshi/pull/3024): Add RegistryUtil class in oshi-core to handle unexpected registry types - [@rohan-coder02](https://github.com/rohan-coder02).
 * [#3040](https://github.com/oshi/oshi/pull/3040): Filter out any USBC powersources on Linux - [@mprins](https://github.com/mprins).
+* [#3041](https://github.com/oshi/oshi/pull/3041): Use systemd as preferred native method for session list - [@dbwiddis](https://github.com/dbwiddis).
 
 # 6.9.0 (2025-09-18), 6.9.1 (2025-10-18)